[feat] 버킷 상세 삭제 적용
삭제 버튼 클릭시 API 연동해서 해당 목표가 삭제되도록 구현
