+++ sidebar/src/components/UI/atoms/read_only_input/style.js
@@ -1,0 +1,12 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    marginTop: theme.spacing(4),
+  },
+  labelRoot: {
+    fontSize: '20px',
+  },
+}));
+
+export default useStyles;
