+++ sidebar/src/components/UI/molecules/detail_list_item/style.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const DetailTextWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+`;
+
+export default DetailTextWrapper;
