+++ sidebar/src/components/UI/organisms/create_detail_list/index.js
@@ -2,18 +2,18 @@ import React from 'react';
 import styled from 'styled-components';
 import { connect } from 'react-redux';
 import DetailAdder from '../../molecules/detail_adder';
-import DetailListItem from '../../molecules/create_detail_list_item';
+import CreateDetailListItem from '../../molecules/create_detail_list_item';
 import { addDetailAction, removeDetailAction } from '../../../../modules/actions/createbucket';
 
 const DetailListWrapper = styled.div`
   padding: 15px;
 `;
 
-const DetailList = ({ details }) => {
+const CreateDetailList = ({ details }) => {
   // TODO : state로 바꿔야함
   const Details = (items) => {
     return items.map((item) => {
-      return <DetailListItem detail={item.title} />;
+      return <CreateDetailListItem detail={item.title} />;
     });
   };
 
@@ -27,4 +27,4 @@ const DetailList = ({ details }) => {
 
 const mapStateToProps = (state) => ({ details: state.createbucket.details });
 
-export default connect(mapStateToProps, { addDetailAction, removeDetailAction })(DetailList);
+export default connect(mapStateToProps, { addDetailAction, removeDetailAction })(CreateDetailList);
