+++ sidebar/src/components/UI/organisms/detail_list/style.js
@@ -1,0 +1,12 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+  list: {
+    width: '100%',
+  },
+  text: {
+    paddingTop: theme.spacing(3),
+  },
+}));
+
+export default useStyles;
