+++ sidebar/src/components/templates/bucket_create_template/index.js
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import { makeStyles } from '@material-ui/core/styles';
 import BucketSearch from '../../UI/organisms/bucket_search';
 import BucketContents from '../../UI/organisms/bucket_contents';
-import DetailList from '../../UI/organisms/create_detail_list';
+import CreateDetailList from '../../UI/organisms/create_detail_list';
 import Buttons from '../../UI/organisms/buttons';
 
 const BucketCreateTemplateWrapper = styled.div`
@@ -31,7 +31,7 @@ const BucketCreateTemplate = () => {
       <BucketCreateTemplateWrapper>
         <BucketSearch />
         <BucketContents />
-        <DetailList />
+        <CreateDetailList />
         <Buttons />
       </BucketCreateTemplateWrapper>
     </>
