+++ sidebar/src/components/templates/my_bucket_list_detail/style.js
@@ -1,5 +1,4 @@
 import { makeStyles } from '@material-ui/core/styles';
-import styled from 'styled-components';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -19,63 +18,6 @@ const useStyles = makeStyles((theme) => ({
     ...theme.mixins.toolbar,
     justifyContent: 'flex-end',
   },
-  title: {
-    flex: 4,
-    '&:hover': {
-      cursor: 'default',
-    },
-  },
-  description: {
-    padding: '8px',
-    '&:hover': {
-      cursor: 'default',
-    },
-  },
-  textField: {
-    padding: '8px',
-  },
-  titleResize: {
-    fontSize: '30px',
-    color: '#555555',
-  },
-  descResize: {
-    fontSize: '20px',
-    color: '#555555',
-  },
-  list: {
-    width: '100%',
-  },
-  text: {
-    paddingTop: theme.spacing(3),
-  },
-  achieveComment: {
-    marginTop: theme.spacing(4),
-  },
-  labelRoot: {
-    fontSize: '20px',
-  },
 }));
 
-const TitleWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  padding: 8px;
-  svg:hover {
-    cursor: pointer;
-  }
-`;
-
-const ButtonWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  padding: 8px;
-  width: 100%;
-  justify-content: flex-end;
-`;
-
-const DetailTextWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-export { useStyles, TitleWrapper, ButtonWrapper, DetailTextWrapper };
+export default useStyles;
